Updated enrollment drop validation to provide frontend with drop eligibility information. Added CanDrop and DropReason properties to EnrollmentDto and implemented CheckCanDropEnrollment method to validate enrollment status and 4-week drop period. Updated GetStudentEnrollmentsAsync to populate drop eligibility. Fixed build errors in AttendanceCheckInValidator, MappingProfile, AttendanceService, and GradeService.

+++ SMARTCAMPUS.BusinessLayer/Concrete/AttendanceService.cs
@@ -242,7 +242,7 @@ namespace SMARTCAMPUS.BusinessLayer.Concrete
                                 checkInDto.Latitude.Value,
                                 checkInDto.Longitude.Value);
 
-                            velocity = (decimal)(distance / 1000.0 / timeDifference); // km/h
+                            velocity = (decimal)((double)distance / 1000.0 / timeDifference); // km/h
 
                             if (velocity > CampusNetworkConstants.MaxRealisticVelocity)
                             {
